move Data and Command types out of specific encoding
